fix referal_system table

+++ console/migrations/m170811_100235_referal_system.php
@@ -17,7 +17,7 @@ class m170811_100235_referal_system extends Migration
             'user_hash' => $this->string(),
             'sum' => $this->float(),
             'invest_name' => $this->string(),
-            'percent' => $this->integer(),
+            'percent' => $this->float(),
         ], $tableOptions);
     }
